Move nodes scale_int.{h,cc} -> characteristic_curve_int_int.{h,cc}
