+++ libspaghetti/source/nodes/values/all.h
@@ -24,9 +24,9 @@
 #ifndef NODES_VALUES_ALL_H
 #define NODES_VALUES_ALL_H
 
+#include "nodes/values/characteristic_curve_int_int.h"
 #include "nodes/values/const_bool.h"
 #include "nodes/values/const_float.h"
 #include "nodes/values/const_int.h"
-#include "nodes/values/scale_int.h"
 
 #endif // NODES_VALUES_ALL_H
