+++ libspaghetti/source/nodes/values/characteristic_curve_int_int.h
